Move convertCurrencyFromExchangeRates to prevent require cycle
