+++ src/components/scenes/TransactionDetailsScene.tsx
@@ -18,12 +18,11 @@ import { useWatch } from '../../hooks/useWatch'
 import { toPercentString } from '../../locales/intl'
 import { lstrings } from '../../locales/strings'
 import { getExchangeDenomination } from '../../selectors/DenominationSelectors'
-import { convertCurrencyFromExchangeRates } from '../../selectors/WalletSelectors'
 import { useSelector } from '../../types/reactRedux'
 import { EdgeSceneProps } from '../../types/routerTypes'
 import { getCurrencyCodeWithAccount } from '../../util/CurrencyInfoHelpers'
 import { matchJson } from '../../util/matchJson'
-import { convertNativeToExchange, darkenHexColor } from '../../util/utils'
+import { convertCurrencyFromExchangeRates, convertNativeToExchange, darkenHexColor } from '../../util/utils'
 import { getMemoTitle } from '../../util/validateMemos'
 import { EdgeAnim } from '../common/EdgeAnim'
 import { SceneWrapper } from '../common/SceneWrapper'
