+++ src/components/tiles/EditableAmountTile.tsx
@@ -3,10 +3,9 @@ import { EdgeCurrencyWallet, EdgeDenomination } from 'edge-core-js'
 import * as React from 'react'
 
 import { lstrings } from '../../locales/strings'
-import { convertCurrencyFromExchangeRates } from '../../selectors/WalletSelectors'
 import { GuiExchangeRates } from '../../types/types'
 import { getWalletFiat } from '../../util/CurrencyWalletHelpers'
-import { DECIMAL_PRECISION, getDenomFromIsoCode, zeroString } from '../../util/utils'
+import { convertCurrencyFromExchangeRates, DECIMAL_PRECISION, getDenomFromIsoCode, zeroString } from '../../util/utils'
 import { EdgeAnim } from '../common/EdgeAnim'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
